+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### BC
+* Renamed `Jwks::getKeyForHeader` to `Jwks::getKeyForJwtHeader`
+
 ### New 
 * It is possible to set list of enabled signature algorithms
+* Check if key from Jwks is intended to use for signing
 
 ## [1.0.0-beta] - 2022-02-08